+++ src/utils/api/index.js
@@ -1,0 +1,2 @@
+export { default as handleAction } from './actions';
+export { default as fetchData } from './fetchData';
