+++ src/utils/format/index.js
@@ -1,0 +1,1 @@
+export { default as addCommas } from './addCommas';
