+++ src/utils/index.js
@@ -1,2 +1,2 @@
-export { default as handleAction } from './actions';
-export { default as fetchData } from './fetchData';
+export * from './api';
+export * from './format';
